requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
certifi==2025.6.15
charset-normalizer==3.4.2
dataclasses-json==0.6.7
distro==1.9.0
frozenlist==1.7.0
googleapis-common-protos==1.70.0
grpcio==1.73.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
idna==3.10
importlib_metadata==8.7.0
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.26
langchain-community==0.3.27
langchain-core==0.3.68
langchain-text-splitters==0.3.8
langsmith==0.4.4
marshmallow==3.26.1
multidict==6.6.3
mypy_extensions==1.1.0
numpy==2.3.1
openai==1.93.0
opentelemetry-api==1.34.1
opentelemetry-distro==0.55b1
opentelemetry-exporter-otlp==1.34.1
opentelemetry-exporter-otlp-proto-common==1.34.1
opentelemetry-exporter-otlp-proto-grpc==1.34.1
opentelemetry-exporter-otlp-proto-http==1.34.1
opentelemetry-instrumentation==0.55b1
opentelemetry-instrumentation-aiohttp-client==0.55b1
opentelemetry-instrumentation-aiohttp-server==0.55b1
opentelemetry-instrumentation-asyncio==0.55b1
opentelemetry-instrumentation-dbapi==0.55b1
opentelemetry-instrumentation-grpc==0.55b1
opentelemetry-instrumentation-httpx==0.55b1
opentelemetry-instrumentation-logging==0.55b1
opentelemetry-instrumentation-openai==0.40.14
opentelemetry-instrumentation-openai-v2==2.1b0
opentelemetry-instrumentation-requests==0.55b1
opentelemetry-instrumentation-sqlalchemy==0.55b1
opentelemetry-instrumentation-sqlite3==0.55b1
opentelemetry-instrumentation-threading==0.55b1
opentelemetry-instrumentation-tortoiseorm==0.55b1
opentelemetry-instrumentation-urllib==0.55b1
opentelemetry-instrumentation-urllib3==0.55b1
opentelemetry-instrumentation-wsgi==0.55b1
opentelemetry-proto==1.34.1
opentelemetry-sdk==1.34.1
opentelemetry-semantic-conventions==0.55b1
opentelemetry-semantic-conventions-ai==0.4.9
opentelemetry-util-http==0.55b1
orjson==3.10.18
packaging==24.2
propcache==0.3.2
protobuf==5.29.5
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
python-dotenv==1.1.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
requests-toolbelt==1.0.0
sniffio==1.3.1
SQLAlchemy==2.0.41
tenacity==9.1.2
tiktoken==0.9.0
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
wrapt==1.17.2
yarl==1.20.1
zipp==3.23.0
zstandard==0.23.0
opentelemetry-instrumentation-asyncio==0.55b1
opentelemetry-instrumentation-dbapi==0.55b1
opentelemetry-instrumentation-logging==0.55b1
opentelemetry-instrumentation-sqlite3==0.55b1
opentelemetry-instrumentation-threading==0.55b1
opentelemetry-instrumentation-urllib==0.55b1
opentelemetry-instrumentation-wsgi==0.55b1
opentelemetry-instrumentation-openai-v2
opentelemetry-instrumentation-aiohttp-client==0.55b1
opentelemetry-instrumentation-aiohttp-server==0.55b1
opentelemetry-instrumentation-grpc==0.55b1
opentelemetry-instrumentation-httpx==0.55b1
opentelemetry-instrumentation-requests==0.55b1
opentelemetry-instrumentation-sqlalchemy==0.55b1
opentelemetry-instrumentation-tortoiseorm==0.55b1
opentelemetry-instrumentation-urllib3==0.55b1
